Rename new_execute() to execute()
