+++ agency/detail/executor_traits/multi_agent_execute_returning_default_container.hpp
@@ -40,7 +40,7 @@ typename new_executor_traits<Executor>::template container<
     >::type
   >;
 
-  return new_executor_traits<Executor>::new_execute(ex, f, container_factory<container_type>{}, shape);
+  return new_executor_traits<Executor>::execute(ex, f, container_factory<container_type>{}, shape);
 } // end multi_agent_execute_returning_user_specified_container()
 
 
@@ -61,16 +61,9 @@ typename new_executor_traits<Executor>::template container<
               Function f,
               typename new_executor_traits<Executor>::shape_type shape)
 {
-  using expected_return_type = typename new_executor_traits<Executor>::template container<
-    typename std::result_of<
-      Function(typename new_executor_traits<Executor>::index_type)
-    >::type
-  >;
-
-  using check_for_member_function = detail::new_executor_traits_detail::has_multi_agent_execute_returning_user_specified_container<
+  using check_for_member_function = detail::new_executor_traits_detail::has_multi_agent_execute_returning_default_container<
     Executor,
-    Function,
-    expected_return_type
+    Function
   >;
 
   return detail::new_executor_traits_detail::multi_agent_execute_returning_default_container(check_for_member_function(), ex, f, shape);
