+++ agency/detail/executor_traits/multi_agent_execute_with_shared_inits_returning_user_specified_container.hpp
@@ -11,7 +11,7 @@ namespace detail
 {
 namespace new_executor_traits_detail
 {
-namespace multi_agent_execute_with_shared_inits_returning_user_specified_container_implementation_strategies
+namespace new_multi_agent_execute_with_shared_inits_returning_user_specified_container_implementation_strategies
 {
 
 
@@ -20,13 +20,13 @@ struct use_multi_agent_execute_with_shared_inits_returning_user_specified_contai
 
 using use_strategy_1 = use_multi_agent_execute_with_shared_inits_returning_user_specified_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
-using has_strategy_1 = has_multi_agent_execute_with_shared_inits_returning_user_specified_container<Container,Executor,Function,Factories...>;
+template<class Executor, class Function, class Factory, class... Factories>
+using has_strategy_1 = has_multi_agent_execute_with_shared_inits_returning_user_specified_container<Executor,Function,Factory,Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_1_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_execute_with_shared_inits_returning_user_specified_container<Container,Executor,Function,Types...>;
+  using type = has_multi_agent_execute_with_shared_inits_returning_user_specified_container<Executor,Function,Factory,Factories...>;
 };
 
 
@@ -35,8 +35,8 @@ struct use_multi_agent_execute_returning_user_specified_container_member_functio
 
 using use_strategy_2 = use_multi_agent_execute_returning_user_specified_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
-using has_strategy_2 = has_multi_agent_execute_returning_user_specified_container<Container,Executor,test_function_returning_int>;
+template<class Executor, class Function, class Factory, class... Factories>
+using has_strategy_2 = has_multi_agent_execute_returning_user_specified_container<Executor,test_function_returning_int,Factory>;
 
 
 // 3.
@@ -44,13 +44,13 @@ struct use_multi_agent_execute_with_shared_inits_returning_void_member_function 
 
 using use_strategy_3 = use_multi_agent_execute_with_shared_inits_returning_void_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_3 = has_multi_agent_execute_with_shared_inits_returning_void<Executor,Function,Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_3_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_execute_with_shared_inits_returning_void<Executor,Function,Types...>;
+  using type = has_multi_agent_execute_with_shared_inits_returning_void<Executor,Function,Factories...>;
 };
 
 
@@ -59,7 +59,7 @@ struct use_multi_agent_execute_returning_void_member_function {};
 
 using use_strategy_4 = use_multi_agent_execute_returning_void_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_4 = has_multi_agent_execute_returning_void<Executor>;
 
 
@@ -68,13 +68,13 @@ struct use_multi_agent_async_execute_with_shared_inits_returning_user_specified_
 
 using use_strategy_5 = use_multi_agent_async_execute_with_shared_inits_returning_user_specified_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
-using has_strategy_5 = has_multi_agent_async_execute_with_shared_inits_returning_user_specified_container<Container,Executor,Function,Factories...>;
+template<class Executor, class Function, class Factory, class... Factories>
+using has_strategy_5 = new_has_multi_agent_async_execute_with_shared_inits_returning_user_specified_container<Executor,Function,Factory,Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_5_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_async_execute_with_shared_inits_returning_user_specified_container<Container,Executor,Function,Types...>;
+  using type = new_has_multi_agent_async_execute_with_shared_inits_returning_user_specified_container<Executor,Function,Factory,Factories...>;
 };
 
 
@@ -83,13 +83,13 @@ struct use_multi_agent_then_execute_with_shared_inits_returning_user_specified_c
 
 using use_strategy_6 = use_multi_agent_then_execute_with_shared_inits_returning_user_specified_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
-using has_strategy_6 = has_multi_agent_then_execute_with_shared_inits_returning_user_specified_container<Container,Executor,Function,typename new_executor_traits<Executor>::template future<void>, Factories...>;
+template<class Executor, class Function, class Factory, class... Factories>
+using has_strategy_6 = new_has_multi_agent_then_execute_with_shared_inits_returning_user_specified_container<Executor,Function,Factory,typename new_executor_traits<Executor>::template future<void>, Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_6_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_then_execute_with_shared_inits_returning_user_specified_container<Container,Executor,Function,typename new_executor_traits<Executor>::template future<void>, Types...>;
+  using type = new_has_multi_agent_then_execute_with_shared_inits_returning_user_specified_container<Executor,Function,Factory,typename new_executor_traits<Executor>::template future<void>, Factories...>;
 };
 
 
@@ -98,13 +98,17 @@ struct use_multi_agent_when_all_execute_and_select_with_shared_inits_member_func
 
 using use_strategy_7 = use_multi_agent_when_all_execute_and_select_with_shared_inits_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_7 = has_multi_agent_when_all_execute_and_select_with_shared_inits<
   detail::index_sequence<0>,
   Executor,
   test_function_returning_void,
   detail::tuple<
-    typename new_executor_traits<Executor>::template future<Container>
+    typename new_executor_traits<Executor>::template future<
+      typename std::result_of<
+        Factory(typename new_executor_traits<Executor>::shape_type)
+      >::type
+    >
   >,
   detail::type_list<Factories...>
 >;
@@ -115,8 +119,8 @@ struct use_multi_agent_async_execute_returning_user_specified_container_member_f
 
 using use_strategy_8 = use_multi_agent_async_execute_returning_user_specified_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
-using has_strategy_8 = has_multi_agent_async_execute_returning_user_specified_container<Container, Executor, Function>;
+template<class Executor, class Function, class Factory, class... Factories>
+using has_strategy_8 = new_has_multi_agent_async_execute_returning_user_specified_container<Executor, Function, Factory>;
 
 
 // 9.
@@ -124,8 +128,8 @@ struct use_multi_agent_then_execute_returning_user_specified_container_member_fu
 
 using use_strategy_9 = use_multi_agent_then_execute_returning_user_specified_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
-using has_strategy_9 = has_multi_agent_then_execute_returning_user_specified_container<Container, Executor, Function, typename new_executor_traits<Executor>::template future<void>>;
+template<class Executor, class Function, class Factory, class... Factories>
+using has_strategy_9 = new_has_multi_agent_then_execute_returning_user_specified_container<Executor, Function, Factory, typename new_executor_traits<Executor>::template future<void>>;
 
 
 // 10.
@@ -133,12 +137,16 @@ struct use_multi_agent_when_all_execute_and_select_member_function {};
 
 using use_strategy_10 = use_multi_agent_when_all_execute_and_select_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_10 = has_multi_agent_when_all_execute_and_select<
   Executor,
   test_function_returning_void,
   detail::tuple<
-    typename new_executor_traits<Executor>::template future<Container>
+    typename new_executor_traits<Executor>::template future<
+      typename std::result_of<
+        Factory(typename new_executor_traits<Executor>::shape_type)
+      >::type
+    >
   >,
   0
 >;
@@ -149,13 +157,13 @@ struct use_multi_agent_async_execute_with_shared_inits_returning_void_member_fun
 
 using use_strategy_11 = use_multi_agent_async_execute_with_shared_inits_returning_void_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_11 = has_multi_agent_async_execute_with_shared_inits_returning_void<Executor,Function,Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_11_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_async_execute_with_shared_inits_returning_void<Executor,Function,Types...>;
+  using type = has_multi_agent_async_execute_with_shared_inits_returning_void<Executor,Function,Factories...>;
 };
 
 
@@ -164,13 +172,13 @@ struct use_multi_agent_then_execute_with_shared_inits_returning_void_member_func
 
 using use_strategy_12 = use_multi_agent_then_execute_with_shared_inits_returning_void_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_12 = has_multi_agent_then_execute_with_shared_inits_returning_void<Executor,Function,typename new_executor_traits<Executor>::template future<void>,Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_12_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_then_execute_with_shared_inits_returning_void<Executor,Function,typename new_executor_traits<Executor>::template future<void>,Types...>;
+  using type = has_multi_agent_then_execute_with_shared_inits_returning_void<Executor,Function,typename new_executor_traits<Executor>::template future<void>,Factories...>;
 };
 
 
@@ -179,13 +187,13 @@ struct use_multi_agent_execute_with_shared_inits_returning_default_container_mem
 
 using use_strategy_13 = use_multi_agent_execute_with_shared_inits_returning_default_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_13 = has_multi_agent_execute_with_shared_inits_returning_default_container<Executor, Function, Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_13_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_execute_with_shared_inits_returning_default_container<Executor, Function, Types...>;
+  using type = has_multi_agent_execute_with_shared_inits_returning_default_container<Executor, Function, Factories...>;
 };
 
 
@@ -194,13 +202,13 @@ struct use_multi_agent_async_execute_with_shared_inits_returning_default_contain
 
 using use_strategy_14 = use_multi_agent_async_execute_with_shared_inits_returning_default_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_14 = has_multi_agent_async_execute_with_shared_inits_returning_default_container<Executor, Function, Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_14_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_async_execute_with_shared_inits_returning_default_container<Executor, Function, Types...>;
+  using type = has_multi_agent_async_execute_with_shared_inits_returning_default_container<Executor, Function, Factories...>;
 };
 
 
@@ -209,13 +217,13 @@ struct use_multi_agent_then_execute_with_shared_inits_returning_default_containe
 
 using use_strategy_15 = use_multi_agent_then_execute_with_shared_inits_returning_default_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_15 = has_multi_agent_then_execute_with_shared_inits_returning_default_container<Executor, Function, typename new_executor_traits<Executor>::template future<void>, Factories...>;
 
-template<class Executor, class Container, class Function, class... Types>
+template<class Executor, class Function, class Factory, class... Factories>
 struct has_strategy_15_workaround_nvbug_1665745
 {
-  using type = has_multi_agent_then_execute_with_shared_inits_returning_default_container<Executor, Function, typename new_executor_traits<Executor>::template future<void>, Types...>;
+  using type = has_multi_agent_then_execute_with_shared_inits_returning_default_container<Executor, Function, typename new_executor_traits<Executor>::template future<void>, Factories...>;
 };
 
 
@@ -224,7 +232,7 @@ struct use_multi_agent_async_execute_returning_void_member_function {};
 
 using use_strategy_16 = use_multi_agent_async_execute_returning_void_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_16 = has_multi_agent_async_execute_returning_void<Executor>;
 
 
@@ -233,7 +241,7 @@ struct use_multi_agent_then_execute_returning_void_member_function {};
 
 using use_strategy_17 = use_multi_agent_then_execute_returning_void_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_17 = has_multi_agent_then_execute_returning_void<Executor>;
 
 
@@ -242,7 +250,7 @@ struct use_multi_agent_execute_returning_default_container_member_function {};
 
 using use_strategy_18 = use_multi_agent_execute_returning_default_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_18 = has_multi_agent_execute_returning_default_container<Executor>;
 
 
@@ -251,7 +259,7 @@ struct use_multi_agent_async_execute_returning_default_container_member_function
 
 using use_strategy_19 = use_multi_agent_async_execute_returning_default_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_19 = has_multi_agent_async_execute_returning_default_container<Executor>;
 
 
@@ -260,7 +268,7 @@ struct use_multi_agent_then_execute_returning_default_container_member_function 
 
 using use_strategy_20 = use_multi_agent_then_execute_returning_default_container_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_20 = has_multi_agent_then_execute_returning_default_container<Executor>;
 
 
@@ -269,7 +277,7 @@ struct use_single_agent_execute_member_function {};
 
 using use_strategy_21 = use_single_agent_execute_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_21 = has_single_agent_execute<Executor>;
 
 
@@ -278,7 +286,7 @@ struct use_single_agent_async_execute_member_function {};
 
 using use_strategy_22 = use_single_agent_async_execute_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_22 = has_single_agent_async_execute<Executor>;
 
 
@@ -287,7 +295,7 @@ struct use_single_agent_then_execute_member_function {};
 
 using use_strategy_23 = use_single_agent_then_execute_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_23 = has_single_agent_then_execute<Executor>;
 
 
@@ -296,7 +304,7 @@ struct use_single_agent_when_all_execute_and_select_member_function {};
 
 using use_strategy_24 = use_single_agent_when_all_execute_and_select_member_function;
 
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using has_strategy_24 = has_single_agent_when_all_execute_and_select<
   Executor,
   test_function_returning_void,
@@ -314,88 +322,88 @@ using use_strategy_25 = use_bare_for_loop;
 
 
 // XXX find a cleaner way to express this
-template<class Executor, class Container, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 using select_multi_agent_execute_with_shared_inits_returning_user_specified_container_implementation =
   typename std::conditional<
-    //has_strategy_1<Executor,Container,Function,Factories...>::value,
-    has_strategy_1_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+    //has_strategy_1<Executor,Function,Factory,Factories...>::value,
+    has_strategy_1_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
     use_strategy_1,
     typename std::conditional<
-      has_strategy_2<Executor,Container,Function,Factories...>::value,
+      has_strategy_2<Executor,Function,Factory,Factories...>::value,
       use_strategy_2,
       typename std::conditional<
-        //has_strategy_3<Executor,Container,Function,Factories...>::value,
-        has_strategy_3_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+        //has_strategy_3<Executor,Function,Factory,Factories...>::value,
+        has_strategy_3_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
         use_strategy_3,
         typename std::conditional<
-          has_strategy_4<Executor,Container,Function,Factories...>::value,
+          has_strategy_4<Executor,Function,Factory,Factories...>::value,
           use_strategy_4,
           typename std::conditional<
-            //has_strategy_5<Executor,Container,Function,Factories...>::value,
-            has_strategy_5_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+            //has_strategy_5<Executor,Function,Factory,Factories...>::value,
+            has_strategy_5_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
             use_strategy_5,
             typename std::conditional<
-              //has_strategy_6<Executor,Container,Function,Factories...>::value,
-              has_strategy_6_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+              //has_strategy_6<Executor,Function,Factory,Factories...>::value,
+              has_strategy_6_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
               use_strategy_6,
               typename std::conditional<
-                has_strategy_7<Executor,Container,Function,Factories...>::value,
+                has_strategy_7<Executor,Function,Factory,Factories...>::value,
                 use_strategy_7,
                 typename std::conditional<
-                  has_strategy_8<Executor,Container,Function,Factories...>::value,
+                  has_strategy_8<Executor,Function,Factory,Factories...>::value,
                   use_strategy_8,
                   typename std::conditional<
-                    has_strategy_9<Executor,Container,Function,Factories...>::value,
+                    has_strategy_9<Executor,Function,Factory,Factories...>::value,
                     use_strategy_9,
                     typename std::conditional<
-                      has_strategy_10<Executor,Container,Function,Factories...>::value,
+                      has_strategy_10<Executor,Function,Factory,Factories...>::value,
                       use_strategy_10,
                       typename std::conditional<
-                        //has_strategy_11<Executor,Container,Function,Factories...>::value,
-                        has_strategy_11_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+                        //has_strategy_11<Executor,Function,Factory,Factories...>::value,
+                        has_strategy_11_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
                         use_strategy_11,
                         typename std::conditional<
-                          //has_strategy_12<Executor,Container,Function,Factories...>::value,
-                          has_strategy_12_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+                          //has_strategy_12<Executor,Function,Factory,Factories...>::value,
+                          has_strategy_12_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
                           use_strategy_12,
                           typename std::conditional<
-                            //has_strategy_13<Executor,Container,Function,Factories...>::value,
-                            has_strategy_13_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+                            //has_strategy_13<Executor,Container,Function,Factory,Factories...>::value,
+                            has_strategy_13_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
                             use_strategy_13,
                             typename std::conditional<
-                              //has_strategy_14<Executor,Container,Function,Factories...>::value,
-                              has_strategy_14_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+                              //has_strategy_14<Executor,Function,Factory,Factories...>::value,
+                              has_strategy_14_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
                               use_strategy_14,
                               typename std::conditional<
-                                //has_strategy_15<Executor,Container,Function,Factories...>::value,
-                                has_strategy_15_workaround_nvbug_1665745<Executor,Container,Function,Factories...>::type::value,
+                                //has_strategy_15<Executor,Function,Factory,Factories...>::value,
+                                has_strategy_15_workaround_nvbug_1665745<Executor,Function,Factory,Factories...>::type::value,
                                 use_strategy_15,
                                 typename std::conditional<
-                                  has_strategy_16<Executor,Container,Function,Factories...>::value,
+                                  has_strategy_16<Executor,Function,Factory,Factories...>::value,
                                   use_strategy_16,
                                   typename std::conditional<
-                                    has_strategy_17<Executor,Container,Function,Factories...>::value,
+                                    has_strategy_17<Executor,Function,Factory,Factories...>::value,
                                     use_strategy_17,
                                     typename std::conditional<
-                                      has_strategy_18<Executor,Container,Function,Factories...>::value,
+                                      has_strategy_18<Executor,Function,Factory,Factories...>::value,
                                       use_strategy_18,
                                       typename std::conditional<
-                                        has_strategy_19<Executor,Container,Function,Factories...>::value,
+                                        has_strategy_19<Executor,Function,Factory,Factories...>::value,
                                         use_strategy_19,
                                         typename std::conditional<
-                                          has_strategy_20<Executor,Container,Function,Factories...>::value,
+                                          has_strategy_20<Executor,Function,Factory,Factories...>::value,
                                           use_strategy_20,
                                           typename std::conditional<
-                                            has_strategy_21<Executor,Container,Function,Factories...>::value,
+                                            has_strategy_21<Executor,Function,Factory,Factories...>::value,
                                             use_strategy_21,
                                             typename std::conditional<
-                                              has_strategy_22<Executor,Container,Function,Factories...>::value,
+                                              has_strategy_22<Executor,Function,Factory,Factories...>::value,
                                               use_strategy_22,
                                               typename std::conditional<
-                                                has_strategy_23<Executor,Container,Function,Factories...>::value,
+                                                has_strategy_23<Executor,Function,Factory,Factories...>::value,
                                                 use_strategy_23,
                                                 typename std::conditional<
-                                                  has_strategy_24<Executor,Container,Function,Factories...>::value,
+                                                  has_strategy_24<Executor,Function,Factory,Factories...>::value,
                                                   use_strategy_24,
                                                   use_strategy_25
                                                 >::type // 24
@@ -425,15 +433,17 @@ using select_multi_agent_execute_with_shared_inits_returning_user_specified_cont
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_with_shared_inits_returning_user_specified_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_with_shared_inits_returning_user_specified_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
-  return ex.template execute<Container>(f, shape, shared_factories...);
+  return ex.execute(f, result_factory, shape, shared_factories...);
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
@@ -493,13 +503,15 @@ multi_agent_execute_with_shared_inits_functor<Result,Function,Shape,TupleOfConta
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_returning_user_specified_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_returning_user_specified_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -515,7 +527,7 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   // wrap f with a functor to map container elements to shared parameters
   auto g = make_multi_agent_execute_with_shared_inits_functor<result_type>(f, shape, shared_param_containers_tuple);
 
-  return ex.template execute<Container>(g, shape);
+  return ex.execute(g, result_factory, shape);
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
@@ -551,15 +563,17 @@ invoke_and_store_result_to_container<Result,Container,Function> make_invoke_and_
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_with_shared_inits_returning_void_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_with_shared_inits_returning_void_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
-  Container results(shape);
+  auto results = result_factory(shape);
 
   ex.execute(make_invoke_and_store_result_to_container<void>(results,f), shape, shared_factories...);
 
@@ -568,13 +582,15 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_returning_void_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_returning_void_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -591,7 +607,7 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   auto g = make_multi_agent_execute_with_shared_inits_functor<result_type>(f, shape, shared_param_containers_tuple);
 
   // wrap g with a functor to store the result to a container
-  Container results(shape);
+  auto results = result_factory(shape);
   auto h = make_invoke_and_store_result_to_container<void>(results, g);
 
   ex.execute(h, shape);
@@ -601,32 +617,36 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_with_shared_inits_returning_user_specified_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_with_shared_inits_returning_user_specified_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // XXX should go through executor_traits for the get()
-  return ex.template async_execute<Container>(f, shape, shared_factories...).get();
+  return ex.new_async_execute(f, result_factory, shape, shared_factories...).get();
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_with_shared_inits_returning_user_specified_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_with_shared_inits_returning_user_specified_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   auto ready = new_executor_traits<Executor>::template make_ready_future<void>(ex);
 
   // XXX should go through executor_traits for the get()
-  return ex.template then_execute<Container>(f, shape, ready, shared_factories...).get();
+  return ex.new_then_execute(f, result_factory, shape, ready, shared_factories...).get();
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
@@ -646,15 +666,18 @@ struct strategy_7_functor
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_when_all_execute_and_select_with_shared_inits_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_when_all_execute_and_select_with_shared_inits_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
-  auto results = new_executor_traits<Executor>::template make_ready_future<Container>(ex, shape);
+  using container_type = typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type;
+  auto results = new_executor_traits<Executor>::template make_ready_future<container_type>(ex, result_factory(shape));
 
   auto futures = detail::make_tuple(std::move(results));
 
@@ -664,13 +687,15 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_returning_user_specified_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_returning_user_specified_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -687,18 +712,20 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   auto g = make_multi_agent_execute_with_shared_inits_functor<result_type>(f, shape, shared_param_containers_tuple);
 
   // XXX should go through executor_traits for this get()
-  return ex.template async_execute<Container>(g, shape).get();
+  return ex.new_async_execute(g, result_factory, shape).get();
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_returning_user_specified_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_returning_user_specified_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -717,7 +744,7 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   auto ready = new_executor_traits<Executor>::template make_ready_future<void>(ex);
 
   // XXX should go through executor_traits for this get()
-  return ex.template then_execute<Container>(g, shape, ready).get();
+  return ex.new_then_execute(g, result_factory, shape, ready).get();
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
@@ -744,13 +771,15 @@ invoke_and_store_to_second_parameter<Function> make_invoke_and_store_to_second_p
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_when_all_execute_and_select_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_when_all_execute_and_select_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -769,7 +798,8 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   // wrap g with a functor to store g's result to the container passed as the second parameter
   auto h = make_invoke_and_store_to_second_parameter(g);
 
-  auto results = new_executor_traits<Executor>::template make_ready_future<Container>(ex, shape);
+  using container_type = typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type;
+  auto results = new_executor_traits<Executor>::template make_ready_future<container_type>(ex, result_factory(shape));
 
   auto futures = detail::make_tuple(std::move(results));
 
@@ -779,15 +809,17 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_with_shared_inits_returning_void_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_with_shared_inits_returning_void_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
-  Container results(shape);
+  auto results = result_factory(shape);
 
   // XXX should call wait() through executor_traits
   ex.async_execute(make_invoke_and_store_result_to_container<void>(results, f), shape, shared_factories...).wait();
@@ -797,15 +829,17 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_with_shared_inits_returning_void_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_with_shared_inits_returning_void_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
-  Container results(shape);
+  auto results = result_factory(shape);
 
   auto ready = new_executor_traits<Executor>::template make_ready_future<void>(ex);
 
@@ -817,16 +851,18 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_with_shared_inits_returning_default_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_with_shared_inits_returning_default_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 
 {
-  Container results(shape);
+  auto results = result_factory(shape);
 
   // discard the container of results returned by this call
   ex.execute(make_invoke_and_store_result_to_container<empty_type>(results,f), shape, shared_factories...);
@@ -836,15 +872,17 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_with_shared_inits_returning_default_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_with_shared_inits_returning_default_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
-  Container results(shape);
+  auto results = result_factory(shape);
 
   // discard the container of results returned by this call
   // XXX should call wait() through executor_traits
@@ -855,13 +893,15 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_returning_void_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_returning_void_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -877,7 +917,7 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   // wrap f with a functor to map container elements to shared parameters
   auto g = make_multi_agent_execute_with_shared_inits_functor<result_type>(f, shape, shared_param_containers_tuple);
 
-  Container results(shape);
+  auto results = result_factory(shape);
 
   // wrap g with a functor to store g's results to a container
   auto h = make_invoke_and_store_result_to_container<void>(results, g);
@@ -891,13 +931,15 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_returning_void_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_returning_void_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -913,7 +955,7 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   // wrap f with a functor to map container elements to shared parameters
   auto g = make_multi_agent_execute_with_shared_inits_functor<result_type>(f, shape, shared_param_containers_tuple);
 
-  Container results(shape);
+  auto results = result_factory(shape);
 
   // wrap g with a functor to store g's results to a container
   auto h = make_invoke_and_store_result_to_container<void>(results, g);
@@ -929,15 +971,17 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_with_shared_inits_returning_default_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_with_shared_inits_returning_default_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
-  Container results(shape);
+  auto results = result_factory(shape);
 
   auto ready = new_executor_traits<Executor>::template make_ready_future<void>(ex);
 
@@ -950,13 +994,15 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_returning_default_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_execute_returning_default_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -972,7 +1018,7 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   // wrap f with a functor to map container elements to shared parameters
   auto g = make_multi_agent_execute_with_shared_inits_functor<result_type>(f, shape, shared_param_containers_tuple);
 
-  Container results(shape);
+  auto results = result_factory(shape);
 
   // wrap g with a functor to store g's results to a container
   auto h = make_invoke_and_store_result_to_container<empty_type>(results, g);
@@ -985,13 +1031,15 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_returning_default_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_async_execute_returning_default_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -1007,7 +1055,7 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   // wrap f with a functor to map container elements to shared parameters
   auto g = make_multi_agent_execute_with_shared_inits_functor<result_type>(f, shape, shared_param_containers_tuple);
 
-  Container results(shape);
+  auto results = result_factory(shape);
 
   // wrap g with a functor to store g's results to a container
   auto h = make_invoke_and_store_result_to_container<empty_type>(results, g);
@@ -1021,13 +1069,15 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_returning_default_container_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_multi_agent_then_execute_returning_default_container_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   // create a tuple of containers holding a shared parameter for each group
   auto shared_param_containers_tuple = make_tuple_of_shared_parameter_containers(ex, shape, shared_factories...);
@@ -1043,7 +1093,7 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
   // wrap f with a functor to map container elements to shared parameters
   auto g = make_multi_agent_execute_with_shared_inits_functor<result_type>(f, shape, shared_param_containers_tuple);
 
-  Container results(shape);
+  auto results = result_factory(shape);
 
   // wrap g with a functor to store g's results to a container
   auto h = make_invoke_and_store_result_to_container<empty_type>(results, g);
@@ -1058,28 +1108,30 @@ Container multi_agent_execute_with_shared_inits_returning_user_specified_contain
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
-template<class Container, class Function, class Shape, class Factory>
+template<class Function, class Factory1, class Shape, class Factory2>
 struct execute_in_for_loop
 {
   mutable Function f;
+  mutable Factory1 result_factory;
   Shape shape;
-  mutable Factory shared_factory;
+  mutable Factory2 shared_factory;
 
   __agency_hd_warning_disable__
-  template<class Function1, class Shape1, class Factory1>
+  template<class F, class F1, class S, class F2>
   __AGENCY_ANNOTATION
-  execute_in_for_loop(Function1&& f_, Shape1&& shape_, Factory1&& shared_factory_)
-    : f(std::forward<Function1>(f_)),
-      shape(std::forward<Shape1>(shape_)),
-      shared_factory(std::forward<Factory1>(shared_factory_))
+  execute_in_for_loop(F&& f_, F1&& result_factory_, S&& shape_, F2&& shared_factory_)
+    : f(std::forward<F>(f_)),
+      result_factory(std::forward<F1>(result_factory_)),
+      shape(std::forward<S>(shape_)),
+      shared_factory(std::forward<F2>(shared_factory_))
   {}
 
   __agency_hd_warning_disable__
   __AGENCY_ANNOTATION
-  Container operator()() const
+  typename std::result_of<Factory1(Shape)>::type
+    operator()() const
   {
-    Container results(shape);
-
+    auto results = result_factory(shape);
     auto shared_arg = shared_factory();
 
     // XXX generalize to multidimensions
@@ -1092,146 +1144,166 @@ struct execute_in_for_loop
   }
 };
 
-template<class Container, class Function, class Shape, class Factory>
+template<class Function, class Factory1, class Shape, class Factory2>
 __AGENCY_ANNOTATION
-execute_in_for_loop<Container,Function,Shape,Factory> make_execute_in_for_loop(Function f, Shape shape, Factory shared_factory)
+execute_in_for_loop<Function,Factory1,Shape,Factory2> make_execute_in_for_loop(Function f, Factory1 result_factory, Shape shape, Factory2 shared_factory)
 {
-  return execute_in_for_loop<Container,Function,Shape,Factory>(f, shape, shared_factory);
+  return execute_in_for_loop<Function,Factory1,Shape,Factory2>(f, result_factory, shape, shared_factory);
 }
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_single_agent_execute_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_single_agent_execute_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   static_assert(sizeof...(Factories) == 1, "This implementation only makes sense for flat (i.e., single-agent) executors");
 
-  return ex.execute(make_execute_in_for_loop<Container>(f, shape, shared_factories...));
+  return ex.execute(make_execute_in_for_loop(f, result_factory, shape, shared_factories...));
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_single_agent_async_execute_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_single_agent_async_execute_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   static_assert(sizeof...(Factories) == 1, "This implementation only makes sense for flat (i.e., single-agent) executors");
 
   // XXX should call get() through executor_traits
-  return ex.async_execute(make_execute_in_for_loop<Container>(f, shape, shared_factories...)).get();
+  return ex.async_execute(make_execute_in_for_loop(f, result_factory, shape, shared_factories...)).get();
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_single_agent_then_execute_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_single_agent_then_execute_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   static_assert(sizeof...(Factories) == 1, "This implementation only makes sense for flat (i.e., single-agent) executors");
 
   auto ready = new_executor_traits<Executor>::template make_ready_future<void>(ex);
 
   // XXX should call get() through executor_traits
-  return ex.then_execute(make_execute_in_for_loop<Container>(f, shape, shared_factories...), ready).get();
+  return ex.then_execute(make_execute_in_for_loop(f, result_factory, shape, shared_factories...), ready).get();
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
-template<class Container, class Function, class Shape, class Factory>
+template<class Function, class Shape, class Factory>
 struct single_agent_when_all_execute_and_select_functor
 {
-  mutable execute_in_for_loop<Container,Function,Shape,Factory> implementation;
+  mutable Function f;
+  Shape shape;
+  mutable Factory shared_factory;
 
   __agency_hd_warning_disable__
+  template<class Container>
   __AGENCY_ANNOTATION
   void operator()(Container& results) const
   {
-    results = implementation();
+    auto shared_arg = shared_factory();
+
+    // XXX generalize to multidimensions
+    for(size_t idx = 0; idx < shape; ++idx)
+    {
+      results[idx] = agency::invoke(f, idx, shared_arg);
+    }
   }
 };
 
 
-template<class Container, class Function, class Shape, class Factory>
+template<class Function, class Shape, class Factory>
 __AGENCY_ANNOTATION
-single_agent_when_all_execute_and_select_functor<Container,Function,Shape,Factory> make_single_agent_when_all_execute_and_select_functor(Function f, Shape shape, Factory shared_factory)
+single_agent_when_all_execute_and_select_functor<Function,Shape,Factory> make_single_agent_when_all_execute_and_select_functor(Function f, Shape shape, Factory shared_factory)
 {
-  return single_agent_when_all_execute_and_select_functor<Container,Function,Shape,Factory>{
-    make_execute_in_for_loop<Container>(f, shape, shared_factory)
-  };
+  return single_agent_when_all_execute_and_select_functor<Function,Shape,Factory>{f, shape, shared_factory};
 }
 
 
 __agency_hd_warning_disable__
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_single_agent_when_all_execute_and_select_member_function,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_single_agent_when_all_execute_and_select_member_function,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   static_assert(sizeof...(Factories) == 1, "This implementation only makes sense for flat (i.e., single-agent) executors");
 
-  auto results = new_executor_traits<Executor>::template make_ready_future<Container>(ex);
+  using result_type = typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type;
+
+  auto results = new_executor_traits<Executor>::template make_ready_future<result_type>(ex, result_factory(shape));
   auto tuple_of_futures = detail::make_tuple(std::move(results));
 
   // XXX should call get() through executor_traits
-  return ex.template when_all_execute_and_select<0>(make_single_agent_when_all_execute_and_select_functor<Container>(f, shape, shared_factories...), tuple_of_futures).get();
+  return ex.template when_all_execute_and_select<0>(make_single_agent_when_all_execute_and_select_functor(f, shape, shared_factories...), tuple_of_futures).get();
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
-template<class Container, class Executor, class Function, class... Factories>
+template<class Executor, class Function, class Factory, class... Factories>
 __AGENCY_ANNOTATION
-Container multi_agent_execute_with_shared_inits_returning_user_specified_container(use_bare_for_loop,
-                                                                                   Executor& ex,
-                                                                                   Function f,
-                                                                                   typename new_executor_traits<Executor>::shape_type shape,
-                                                                                   Factories... shared_factories)
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type
+  multi_agent_execute_with_shared_inits_returning_user_specified_container(use_bare_for_loop,
+                                                                           Executor& ex,
+                                                                           Function f,
+                                                                           Factory result_factory,
+                                                                           typename new_executor_traits<Executor>::shape_type shape,
+                                                                           Factories... shared_factories)
 {
   static_assert(sizeof...(Factories) == 1, "This implementation only makes sense for flat (i.e., single-agent) executors");
 
-  auto implementation = make_execute_in_for_loop<Container>(f, shape, shared_factories...);
+  auto implementation = make_execute_in_for_loop(f, result_factory, shape, shared_factories...);
   return implementation();
 } // end multi_agent_execute_with_shared_inits_returning_user_specified_container()
 
 
-} // end multi_agent_execute_with_shared_inits_returning_user_specified_container_implementation_strategies
+} // end new_multi_agent_execute_with_shared_inits_returning_user_specified_container_implementation_strategies
 } // end new_executor_traits_detail
 } // end detail
 
 
 template<class Executor>
-  template<class Container, class Function, class... Factories,
+  template<class Function, class Factory, class... Factories,
            class Enable>
 __AGENCY_ANNOTATION
-Container new_executor_traits<Executor>
+typename std::result_of<Factory(typename new_executor_traits<Executor>::shape_type)>::type new_executor_traits<Executor>
   ::execute(typename new_executor_traits<Executor>::executor_type& ex,
             Function f,
+            Factory result_factory,
             typename new_executor_traits<Executor>::shape_type shape,
             Factories... shared_factories)
 {
-  namespace ns = detail::new_executor_traits_detail::multi_agent_execute_with_shared_inits_returning_user_specified_container_implementation_strategies;
+  namespace ns = detail::new_executor_traits_detail::new_multi_agent_execute_with_shared_inits_returning_user_specified_container_implementation_strategies;
 
   using implementation_strategy = ns::select_multi_agent_execute_with_shared_inits_returning_user_specified_container_implementation<
     Executor,
-    Container,
     Function,
+    Factory,
     Factories...
   >;
 
-  return ns::multi_agent_execute_with_shared_inits_returning_user_specified_container<Container>(implementation_strategy(), ex, f, shape, shared_factories...);
+  return ns::multi_agent_execute_with_shared_inits_returning_user_specified_container(implementation_strategy(), ex, f, result_factory, shape, shared_factories...);
 } // end new_executor_traits::execute()
